requirements.txt
click==6.7
crypto==1.4.1
cryptography==2.2.2
django==2.0.2
Flask==1.0.2
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.10
mysql-connector-python==8.0.13
Naked==0.1.31
pymessenger==0.0.7.0
PyMySQL==0.9.1
pyOpenSSL==18.0.0
requests-toolbelt==0.8.0
requirements==0.1
